requirements.txt
vtkplotter
matplotlib
pydicom
Pillow
numpy
